mcp/databricks-executor/requirements.txt
mcp>=1.0.0
databricks-sdk>=0.20.0
pyyaml>=6.0
